// src/main/common/form/models/statementOfMeans/expenses/expenseType.ts
export enum ExpenseType {
  OTHER = 'other expense',
  MORTGAGE = 'mortgage',
  RENT = 'rent',
  MAINTENANCE_PAYMENTS = 'maintenance payments',
  COUNCIL_TAX = 'Council Tax',
  GAS = 'gas',
  ELECTRICITY = 'electricity',
  WATER = 'water',
  TRAVEL = 'travel',
  SCHOOL_COSTS = 'school costs (include clothing)',
  FOOD_HOUSEKEEPING = 'food and housekeeping',
  TV_AND_BROADBAND = 'TV and broadband',
  HIRE_PURCHASES = 'hire purchase',
  MOBILE_PHONE = 'mobile phone',
}